@@ -4,18 +4,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "bufferutil": "^4.0.1",
-    "express": "^4.17.1",
+    "@testing-library/jest-dom": "^6.6.4",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "bufferutil": "^4.0.9",
+    "express": "^5.1.0",
     "path": "^0.12.7",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "utf-8-validate": "^5.0.2",
-    "ws": "^7.2.1"
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-router-dom": "^7.7.1",
+    "react-scripts": "^5.0.1",
+    "utf-8-validate": "^6.0.5",
+    "ws": "^8.18.3"
   },
   "scripts": {
     "predeploy": "npm run build",
@@ -41,6 +41,10 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^2.2.0"
-  }
+    "gh-pages": "^6.3.0"
+  },
+  "engines": {
+  "node": ">=16 <=18"
+}
+
 }
